framework: switch example directory (#6240)
* framework: brief intro about Apollo Cyber RT

More to come with our official release

* framework: add docs about how to add new lidar driver

* framework: switch example directory
