+++ cyber/examples/common_component_example/BUILD
@@ -3,23 +3,23 @@ load("//tools:cpplint.bzl", "cpplint")
 package(default_visibility = ["//visibility:public"])
 
 cc_binary(
-    name = "libtimer_component_sample.so",
+    name = "libcommon_component_example.so",
     linkopts = ["-shared"],
     linkstatic = False,
-    deps = [":timer_component_sample_lib"],
+    deps = [":common_component_example_lib"],
 )
 
 cc_library(
-    name = "timer_component_sample_lib",
+    name = "common_component_example_lib",
     srcs = [
-        "timer_component_sample.cc",
+        "common_component_example.cc",
     ],
     hdrs = [
-        "timer_component_sample.h",
+        "common_component_example.h",
     ],
     deps = [
         "//cyber",
-        "//cyber/samples/proto:samples_cc_proto",
+        "//cyber/examples/proto:examples_cc_proto",
     ],
 )
