+++ cyber/examples/common_component_example/common_component_example.h
@@ -17,21 +17,16 @@
 
 #include "cyber/class_loader/class_loader.h"
 #include "cyber/component/component.h"
-#include "cyber/component/timer_component.h"
-#include "cyber/samples/proto/samples.pb.h"
+#include "cyber/examples/proto/examples.pb.h"
 
-using apollo::cyber::samples::proto::Driver;
+using apollo::cyber::examples::proto::Driver;
 using apollo::cyber::Component;
 using apollo::cyber::ComponentBase;
-using apollo::cyber::TimerComponent;
-using apollo::cyber::Writer;
 
-class TimerComponentSample : public TimerComponent {
+class CommonComponentSample : public Component<Driver, Driver> {
  public:
   bool Init() override;
-  bool Proc() override;
-
- private:
-  std::shared_ptr<Writer<Driver>> driver_writer_ = nullptr;
+  bool Proc(const std::shared_ptr<Driver>& msg0,
+            const std::shared_ptr<Driver>& msg1) override;
 };
-CYBER_REGISTER_COMPONENT(TimerComponentSample)
+CYBER_REGISTER_COMPONENT(CommonComponentSample)
