+++ cyber/examples/talker.cc
@@ -15,23 +15,32 @@
  *****************************************************************************/
 
 #include "cyber/cyber.h"
-#include "cyber/samples/proto/samples.pb.h"
+#include "cyber/examples/proto/examples.pb.h"
+#include "cyber/time/rate.h"
+#include "cyber/time/time.h"
 
-void MessageCallback(
-    const std::shared_ptr<apollo::cyber::samples::proto::Chatter>& msg) {
-  AINFO << "Received message seq-> " << msg->seq();
-  AINFO << "msgcontent->" << msg->content();
-}
+using apollo::cyber::Rate;
+using apollo::cyber::Time;
+using apollo::cyber::examples::proto::Chatter;
 
-int main(int argc, char* argv[]) {
+int main(int argc, char *argv[]) {
   // init cyber framework
   apollo::cyber::Init(argv[0]);
-  // create listener node
-  auto listener_node = apollo::cyber::CreateNode("listener");
-  // create listener
-  auto listener =
-      listener_node->CreateReader<apollo::cyber::samples::proto::Chatter>(
-          "channel/chatter", MessageCallback);
-  apollo::cyber::WaitForShutdown();
+  // create talker node
+  auto talker_node = apollo::cyber::CreateNode("talker");
+  // create talker
+  auto talker = talker_node->CreateWriter<Chatter>("channel/chatter");
+  Rate rate(1.0);
+  while (apollo::cyber::OK()) {
+    static uint64_t seq = 0;
+    auto msg = std::make_shared<Chatter>();
+    msg->set_timestamp(Time::Now().ToNanosecond());
+    msg->set_lidar_timestamp(Time::Now().ToNanosecond());
+    msg->set_seq(seq++);
+    msg->set_content("Hello, apollo!");
+    talker->Write(msg);
+    AINFO << "talker sent a message!";
+    rate.Sleep();
+  }
   return 0;
 }
