+++ cyber/examples/timer_component_example/timer.dag
@@ -1,5 +1,5 @@
 module_config {
-    module_library : "/apollo/bazel-bin/cyber/samples/timer_component_sample/libtimer_component_sample.so"
+    module_library : "/apollo/bazel-bin/cyber/examples/timer_component_example/libtimer_component_example.so"
     timer_components {
         class_name : "TimerComponentSample"
         config {
