+++ cyber/examples/timer_component_example/timer.launch
@@ -1,7 +1,7 @@
 <cyber>
     <module>
         <name>timer</name>
-        <dag_conf>/apollo/cyber/samples/timer_component_sample/timer.dag</dag_conf>
+        <dag_conf>/apollo/cyber/examples/timer_component_example/timer.dag</dag_conf>
         <process_name>timer</process_name>
     </module>
 </cyber>
