+++ cyber/examples/timer_component_example/timer_component_example.cc
@@ -14,11 +14,11 @@
  * limitations under the License.
  *****************************************************************************/
 
-#include "cyber/samples/timer_component_sample/timer_component_sample.h"
+#include "cyber/examples/timer_component_example/timer_component_example.h"
 
 #include "cyber/class_loader/class_loader.h"
 #include "cyber/component/component.h"
-#include "cyber/samples/proto/samples.pb.h"
+#include "cyber/examples/proto/examples.pb.h"
 
 bool TimerComponentSample::Init() {
   driver_writer_ = node_->CreateWriter<Driver>("/carstatus/channel");
@@ -30,7 +30,7 @@ bool TimerComponentSample::Proc() {
   auto out_msg = std::make_shared<Driver>();
   out_msg->set_msg_id(i++);
   driver_writer_->Write(out_msg);
-  AINFO << "timer_component_sample: Write drivermsg->"
+  AINFO << "timer_component_example: Write drivermsg->"
         << out_msg->ShortDebugString();
   return true;
 }
