+++ cyber/examples/timer_component_example/timer_component_example.h
@@ -13,19 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *****************************************************************************/
-#include "cyber/samples/common_component_sample/common_component_sample.h"
+#include <memory>
 
 #include "cyber/class_loader/class_loader.h"
 #include "cyber/component/component.h"
+#include "cyber/component/timer_component.h"
+#include "cyber/examples/proto/examples.pb.h"
 
-bool CommonComponentSample::Init() {
-  AINFO << "Commontest component init";
-  return true;
-}
+using apollo::cyber::examples::proto::Driver;
+using apollo::cyber::Component;
+using apollo::cyber::ComponentBase;
+using apollo::cyber::TimerComponent;
+using apollo::cyber::Writer;
 
-bool CommonComponentSample::Proc(const std::shared_ptr<Driver>& msg0,
-                               const std::shared_ptr<Driver>& msg1) {
-  AINFO << "Start common component Proc [" << msg0->msg_id() << "] ["
-        << msg1->msg_id() << "]";
-  return true;
-}
+class TimerComponentSample : public TimerComponent {
+ public:
+  bool Init() override;
+  bool Proc() override;
+
+ private:
+  std::shared_ptr<Writer<Driver>> driver_writer_ = nullptr;
+};
+CYBER_REGISTER_COMPONENT(TimerComponentSample)
